Make error handling better in roll()
Before you could attempt to roll, even if there was invalid input
in dice expression entry, because validating function is called
on key-release-event and if roll button was pressed very quickly,
a button press on a roll button could get registered. In that case
a segfault happened.
The same still applies, except the segfault, because the dice
expression result is checked in roll() for errors.

Print error messages on integer overflow(only in dice expression)
and on syntax error.

Add abort() on out of memory.

Refactor roll().