Use prefix to config paths in Makefile
Some resources are needed and their paths too when running and
installing the program. Use prefix variable in Makefile to set
the paths for these resources. config.h file will be created from
the config.g.orig template and resources will be copied to the
wanted directories.

Add possibility to compile with debug parameters.