Add install and uninstall goals
Compile and install with:
        make [PREFIX=<absolute_path>]
        make install

And uninstall:
        make uninstall

Install and uninstall goals read $PREFIX from a $prefix_file.

To run the program, you need to make install too. If you do only
make, the resources are in the wrong directory. Install copies
them.

I added the install goal to be able to just copy the files, if you
install the program to $PREFIX, to which higher privileges are
needed(ie. root). Compiling as root wouldn't be nice.