Disable resizing the window
Remove the fullscreen menuitem and the possibility to resize the
window.
Currently, if the window is made wider, the width of some labels
and entries of spinbuttons will not be uniform. Some stays the
same and others grow wider. If the height of the window is grown,
there will blank area on the bottom of the window, no height of
any other widget will grow.

If resizing the window is to be enabled again, the width of similar
widgets should stay the same and the size of the textview should
change.